Refactor: move api call to get products to DataContext move api call to get wishlist to WishlistContext move api call to get cart to CartContext set the UserDetails before the isUserLogin in the AuthContext since token was not available in the CartContext & WishlistContext when user isUserLogin
