+++ src/App.js
@@ -11,88 +11,9 @@ import SignUp from "./pages/SignUp";
 import Wishlist from "./pages/Wishlist";
 import Cart from "./pages/Cart";
 import PrivateRoutes from "./utils/PrivateRoutes";
-import { DataContext } from "./contexts/DataContext";
-import { AuthContext } from "./contexts/AuthContext";
-import { WishlistContext } from "./contexts/WishlistContext";
-import { CartContext } from "./contexts/CartContext";
 import Footer from "./components/Footer";
 
 function App() {
-    const { setData } = useContext(DataContext);
-    const { isUserLogin, token, userId } = useContext(AuthContext);
-    const { dispatch: wishlistDispatch } = useContext(WishlistContext);
-    const { dispatch: cartDispatch } = useContext(CartContext);
-
-    // get data from the server and save them in context
-    const getData = async () => {
-        try {
-            const response = await axios.get("/products");
-            if (response.data.success) {
-                setData(response.data.product);
-            }
-        } catch (err) {
-            console.log(err);
-        }
-    };
-
-    // get cart data from server and save them in context
-    const getCart = async () => {
-        try {
-            const response = await axios.get(`/cart/${userId}`, {
-                headers: {
-                    authorization: token,
-                },
-            });
-            if (response.data.success) {
-                cartDispatch({
-                    type: "SYNC_CART",
-                    payload: {
-                        product: response.data.cart.cartItems,
-                    },
-                });
-            }
-        } catch (error) {
-            console.log(error);
-        }
-    };
-
-    // get wishlist data from the server and save them in context
-    const getWishlist = async () => {
-        try {
-            const response = await axios.get(`/wishlist/${userId}`, {
-                headers: {
-                    authorization: token,
-                },
-            });
-            if (response.data.success) {
-                wishlistDispatch({
-                    type: "SYNC_WISHLIST",
-                    payload: {
-                        product: response.data.wishlist.wishlistItems,
-                    },
-                });
-            }
-        } catch (error) {
-            console.log(error);
-        }
-    };
-
-    useEffect(() => {
-        getData();
-        if (isUserLogin) {
-            getWishlist();
-            getCart();
-        }
-    }, []);
-
-    useEffect(() => {
-        console.log("user looged");
-        if (isUserLogin) {
-            getWishlist();
-            getCart();
-        }
-    }, [isUserLogin]);
-
     return (
         <div className="App">
             <header className="App-header">
